backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
redis>=5.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
slowapi>=0.1.8
alembic>=1.12.0
openai>=1.0.0
aliyun-python-sdk-core>=2.13.3
